telegram_opusmt_bot/requirements.txt
pycld2
websocket
iso639-lang

